ci fix

+++ incl/core/consts.h
@@ -2,5 +2,6 @@
 #include "core/config.h"
 
 //Consts
-const f64 pi=acosl(-1),eps=FP_EPS;
+// const f64 pi=numbers::pi_v<f64>,eps=FP_EPS;
+const f64 pi=acos(-1),eps=FP_EPS;
 const int dir[4][2]={{1,0},{0,1},{-1,0},{0,-1}};
